4/5Ola

+++ App.Domain/Entities/Building.cs
@@ -17,7 +17,7 @@ namespace Domain.App.Entities
 
         public decimal TotalSurfaceArea { get; set; }
 
-        public Nullable<byte> NumberOfFloor { get; set; }
+        public Nullable<int> NumberOfFloor { get; set; }
 
         public string LicenseNumber { get; set; }
         public string StampingNumber { get; set; }
